+++ src/auth_server/authServer.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 	"crud/src/authentication"
-	"crud/src/bookingimpl"
+	"crud/src/bookingimpl/dao"
 	"fmt"
 	"log"
 	"net"
@@ -17,7 +17,7 @@ type server struct {
 
 func (*server) Authenticate(ctx context.Context, input *authentication.Credentials) (*authentication.Result, error) {
 	fmt.Println("In grpcServer : Authenticate start")
-	d := bookingimpl.NewBookingDao()
+	d := dao.NewBookingDao()
 	res, err := d.AuthenticateUser(input)
 	fmt.Println("In grpcServer : Authenticate stop")
 	return &authentication.Result{
