+++ src/bookingimpl/rest/rest_test.go
@@ -1,4 +1,4 @@
-package bookingimpl
+package rest
 
 import (
 	"crud/src/domain"
